==> 00-01- ATM GTFS.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt

# Load the data
routes_df = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\Lombardia GTFS\ATM-UTF8\routes.txt")
trips_df = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\Lombardia GTFS\ATM-UTF8\trips.txt")
stop_times_df = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\Lombardia GTFS\ATM-UTF8\stop_times.txt")
stops_df = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\Lombardia GTFS\ATM-UTF8\stops.txt")

## BUS

routes_df.tail(10)

bus_routes = routes_df[routes_df["route_type"] == 3]

bus_routes

bus_trips = trips_df[trips_df["route_id"].isin(bus_routes['route_id'])]

bus_trips

bus_stop_time=stop_times_df[stop_times_df['trip_id'].isin(bus_trips['trip_id'])]

bus_stop_time

bus_stops=bus_stop_time[bus_stop_time['stop_id'].isin(stops_df['stop_id'])]

bus_stop=stops_df[stops_df['stop_id'].isin(bus_stops['stop_id'])]

#s

s.drop_duplicates(subset='stop_id', keep='last')

s.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\bus.csv")

## METRO

metro_routes = routes_df[routes_df["route_type"] == 1]

metro_trips = trips_df[trips_df["route_id"].isin(metro_routes['route_id'])]

metro_stop_time=stop_times_df[stop_times_df['trip_id'].isin(metro_trips['trip_id'])]

metro_stops= stops_df[stops_df['stop_id'].isin(metro_stop_time['stop_id'])]

metro_stops.drop_duplicates(subset='stop_id', keep='last')

metro_stops.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\metro.csv")

## TRAM

tram_routes = routes_df[routes_df["route_type"] == 0]

tram_trips= trips_df[trips_df["route_id"].isin(tram_routes['route_id'])]

tram_stop_time=stop_times_df[stop_times_df['trip_id'].isin(tram_trips['trip_id'])]

tram_stops= stops_df[stops_df["stop_id"].isin(tram_stop_time['stop_id'])]

tram_stops.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\tram.csv")

## BUS NOT IN LINE 90-91

novanta=pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-22-01-2024\90.csv")

s = s[~s['stop_id'].isin(novanta['stop_id'])]
s

s.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\bus-not-90.csv")

## BUS STOPS OF 90-91

s = bus_stop[bus_stop['stop_id'].isin(novanta['stop_id'])]
s

s.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\bus90.csv")
