==> 01-ABC-mode.ipynb <==
# Re adaptation of the ABC location Policy for primarily classifying the Mulimodality conditions in Milan's PUMPS Area.

### the model runs in a parts, so you need to run each model, restart Kernal before moving to the other one.

import geopandas as gpd
from shapely.geometry import Point
import pandas as pd

# Load the shapefiles
train_gdf = gpd.read_file(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\train_stops_PUMS.shp").to_crs(epsg=32632)
metro_gdf = gpd.read_file(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\metro_stops_PUMS.shp").to_crs(epsg=32632)
bus_gdf = gpd.read_file(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\bus_90.shp").to_crs(epsg=32632)
tram_gdf = gpd.read_file(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\tram_pumS.shp").to_crs(epsg=32632)
parking_gdf=gpd.read_file(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD Readings\ABC\B\P+R.shp").to_crs(epsg=32632)
reg_bus_gdf=gpd.read_file(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\reg-bus.shp").to_crs(epsg=32632)
bus_tranne_90= gpd.read_file(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\bus_not_90.shp").to_crs(epsg=32632)
highway_exit= gpd.read_file(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\highway_exit.shp").to_crs(epsg=32632)
parking= gpd.read_file(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\parking_lot.shp").to_crs(epsg=32632)
#bus_others_gdf=gpd.read_file(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-22-01-2024\bus-others.shp").to_crs(epsg=32632)


# Adding a column 'int_scr' with the specified values for each DataFrame
train_gdf['int_scr'] = 6
metro_gdf['int_scr'] = 5
tram_gdf['int_scr'] = 4
bus_gdf['int_scr'] = 3
bus_tranne_90['int_scr'] = 3
reg_bus_gdf['int_scr'] = 2
parking_gdf['int_scr'] = 1
highway_exit['int_scr'] = 1
parking['int_scr'] = 1

## Train_ABC-model

# Merge the point layers into a single GeoDataFrame
points_gdf = gpd.GeoDataFrame(pd.concat([metro_gdf, bus_gdf, tram_gdf,parking_gdf,reg_bus_gdf,bus_tranne_90,highway_exit,parking], ignore_index=True))

# Create a 125-meter buffer around the Train points
buffer_distance = 125
train_gdf['buffer'] = train_gdf.geometry.buffer(buffer_distance)

train_scores = {}

# Perform the intersection and calculate the sum of unique 'int_scr' values for each buffered train feature
for index, train_feature in train_gdf.iterrows():
    buffer_zone = train_feature['buffer']
    # Find points within the buffer zone
    intersected_points = points_gdf[points_gdf.geometry.intersects(buffer_zone)]
    # Sum the unique 'int_scr' values
    unique_scores = set(intersected_points['int_scr'].tolist())
    # Update train_gdf with the unique sum score for each train feature
    train_gdf.at[index, 'unique_sum'] = sum(unique_scores)

# Now train_gdf has an additional column 'unique_sum' with the calculated scores

train_gdf.at[index,'unique_sum'] = int(sum(unique_scores))
train_gdf['scr_abc']= train_gdf['unique_sum'] + train_gdf ['int_scr']
#train_gdf

train_gdf.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\train_scr.csv")

## METRO ABC MODEL

# Merge the point layers into a single GeoDataFrame
per_metro_gdf = gpd.GeoDataFrame(pd.concat([train_gdf, bus_gdf, tram_gdf,parking_gdf,reg_bus_gdf,bus_tranne_90,highway_exit,parking], ignore_index=True))

# Create a 125-meter buffer around the Train points
buffer_distance = 125
metro_gdf['buffer'] = metro_gdf.geometry.buffer(buffer_distance)

metro_scores = {}

# Perform the intersection and calculate the sum of unique 'int_scr' values for each buffered train feature
for index, metro_feature in metro_gdf.iterrows():
    buffer_zone = metro_feature['buffer']
    # Find points within the buffer zone
    intersected_points = per_metro_gdf[per_metro_gdf.geometry.intersects(buffer_zone)]
    # Sum the unique 'int_scr' values
    unique_scores = set(intersected_points['int_scr'].tolist())
    metro_scores[metro_feature['stop_id']] = sum(unique_scores)

# Perform the intersection and calculate the sum of unique 'int_scr' values for each buffered train feature
for index, metro_feature in metro_gdf.iterrows():
    buffer_zone = metro_feature['buffer']
    # Find points within the buffer zone
    intersected_points = per_metro_gdf[per_metro_gdf.geometry.intersects(buffer_zone)]
    # Sum the unique 'int_scr' values
    unique_scores = set(intersected_points['int_scr'].tolist())
    # Update train_gdf with the unique sum score for each train feature
    metro_gdf.at[index, 'unique_sum'] = sum(unique_scores)

# Now metro_gdf has an additional column 'unique_sum' with the calculated scores


metro_gdf.at[index, 'unique_sum'] = int(sum(unique_scores))
metro_gdf['scr_abc']= metro_gdf['unique_sum'] + metro_gdf ['int_scr']
metro_gdf

metro_gdf.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\metro_scr.csv")

## TRAM_ABC_MODEL

# Merge the point layers into a single GeoDataFrame
per_tram_gdf = gpd.GeoDataFrame(pd.concat([metro_gdf, bus_gdf, train_gdf,parking_gdf,reg_bus_gdf,bus_tranne_90,highway_exit,parking], ignore_index=True))

# Create a 125-meter buffer around the Train points
buffer_distance = 125
tram_gdf['buffer'] = tram_gdf.geometry.buffer(buffer_distance)

tram_gdf

tram_scores = {}

# Perform the intersection and calculate the sum of unique 'int_scr' values for each buffered train feature
for index, tram_feature in tram_gdf.iterrows():
    buffer_zone = tram_feature['buffer']
    # Find points within the buffer zone
    intersected_points = per_tram_gdf[per_tram_gdf.geometry.intersects(buffer_zone)]
    # Sum the unique 'int_scr' values
    unique_scores = set(intersected_points['int_scr'].tolist())
    tram_scores[tram_feature['stop_id']] = sum(unique_scores)

# Perform the intersection and calculate the sum of unique 'int_scr' values for each buffered train feature
for index, tram_feature in tram_gdf.iterrows():
    buffer_zone = tram_feature['buffer']
    # Find points within the buffer zone
    intersected_points = per_tram_gdf[per_tram_gdf.geometry.intersects(buffer_zone)]
    # Sum the unique 'int_scr' values
    unique_scores = set(intersected_points['int_scr'].tolist())
    # Update train_gdf with the unique sum score for each train feature
    tram_gdf.at[index, 'unique_sum'] = sum(unique_scores)

# Now tram_gdf has an additional column 'unique_sum' with the calculated scores


tram_gdf.at[index, 'unique_sum'] = int(sum(unique_scores))
tram_gdf['scr_abc']= tram_gdf['unique_sum'] + tram_gdf['int_scr']
tram_gdf

tram_gdf.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\tram_scr.csv")

## BUS_ABC_MODEL

# Merge the point layers into a single GeoDataFrame
per_bus_gdf = gpd.GeoDataFrame(pd.concat([metro_gdf, train_gdf, tram_gdf,parking_gdf,reg_bus_gdf,bus_tranne_90,highway_exit,parking], ignore_index=True))

# Create a 125-meter buffer around the Train points
buffer_distance = 125
bus_gdf['buffer'] = bus_gdf.geometry.buffer(buffer_distance)

bus_scores = {}

# Perform the intersection and calculate the sum of unique 'int_scr' values for each buffered train feature
for index, bus_feature in bus_gdf.iterrows():
    buffer_zone = bus_feature['buffer']
    # Find points within the buffer zone
    intersected_points = per_bus_gdf[per_bus_gdf.geometry.intersects(buffer_zone)]
    # Sum the unique 'int_scr' values
    unique_scores = set(intersected_points['int_scr'].tolist())
    bus_scores[bus_feature['stop_id']] = sum(unique_scores)

# Perform the intersection and calculate the sum of unique 'int_scr' values for each buffered train feature
for index, bus_feature in bus_gdf.iterrows():
    buffer_zone = bus_feature['buffer']
    # Find points within the buffer zone
    intersected_points = per_bus_gdf[per_bus_gdf.geometry.intersects(buffer_zone)]
    # Sum the unique 'int_scr' values
    unique_scores = set(intersected_points['int_scr'].tolist())
    # Update train_gdf with the unique sum score for each train feature
    bus_gdf.at[index, 'unique_sum'] = sum(unique_scores)

# Now tram_gdf has an additional column 'unique_sum' with the calculated scores

bus_gdf.at[index, 'unique_sum'] = int(sum(unique_scores))
bus_gdf['scr_abc']= bus_gdf['unique_sum'] + bus_gdf['int_scr']
bus_gdf

bus_gdf.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\bus_scr.csv")
