==> 02-00 Scoring Mechnanisms.ipynb <==
import geopandas as gpd
from shapely.geometry import Point
import pandas as pd

## 1. Loading Scores from 00

# Load the dataset
train_scr = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\train_scr.csv")
metro_scr = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\metro_scr.csv")
tram_scr = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\tram_scr.csv")
bus_scr = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\bus_scr.csv")

## 2. Assining categories 

### 2.1 Train

gdf=train_scr

# Define the assign_
def assign_cluster(score, low_threshold, high_threshold):
    if score <= low_threshold:
        return 'Low'
    elif score > low_threshold and score <= high_threshold:
        return 'Medium'
    else:
        return 'High'

# Assuming the 'catch_area' column exists and is already calculated as 'scr_abc' * 35
gdf['catch_area'] = gdf['scr_abc'] * 35

# Calculate the thresholds for clustering
low_threshold = gdf['scr_abc'].quantile(0.25)
high_threshold = gdf['scr_abc'].quantile(0.75)

# Apply the function to create a new column 'Cluster'
gdf['Cluster'] = gdf['scr_abc'].apply(lambda x: assign_cluster(x, low_threshold, high_threshold))

# Find the maximum 'catch_area' for each cluster
max_catch_area_per_cluster = gdf.groupby('Cluster')['catch_area'].transform('max')

# Assign the max 'catch_area' to all members of the same cluster
gdf['catch_area'] = max_catch_area_per_cluster
gdf = gpd.GeoDataFrame(gdf, geometry=gpd.points_from_xy(gdf.s_lon, gdf.s_lat))
# Set the original CRS to WGS84 (lat/lon)
gdf.crs = 'EPSG:4326'  # WGS 84
# Reproject the GeoDataFrame to WGS 84 / UTM zone 32N
gdf = gdf.to_crs(epsg=32632)
# Select the fields you want to include in the shapefile
fields = ['stop_id', 'stop_name', 'stop_desc', 'pums', 's_lat', 's_lon', 'int_scr', 'scr_abc','Cluster', 'catch_area']
gdf = gdf[fields]
gdf.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\02-Catchment_area\02-train_catchment.csv")

# WHEN YOU FINISH EXPORTING TRAIN, RESTERT KERNAL AND RUN ONLY TILL 1. Loading Scores from 00

### 2.2 Metro

gdf=metro_scr

# Define the assign_
def assign_cluster(score, low_threshold, high_threshold):
    if score <= low_threshold:
        return 'Low'
    elif score > low_threshold and score <= high_threshold:
        return 'Medium'
    else:
        return 'High'

# Assuming the 'catch_area' column exists and is already calculated as 'scr_abc' * 35
gdf['catch_area'] = gdf['scr_abc'] * 35

# Calculate the thresholds for clustering
low_threshold = gdf['scr_abc'].quantile(0.25)
high_threshold = gdf['scr_abc'].quantile(0.75)

# Apply the function to create a new column 'Cluster'
gdf['Cluster'] = gdf['scr_abc'].apply(lambda x: assign_cluster(x, low_threshold, high_threshold))

# Find the maximum 'catch_area' for each cluster
max_catch_area_per_cluster = gdf.groupby('Cluster')['catch_area'].transform('max')

# Assign the max 'catch_area' to all members of the same cluster
gdf['catch_area'] = max_catch_area_per_cluster
gdf = gpd.GeoDataFrame(gdf, geometry=gpd.points_from_xy(gdf.s_lon, gdf.s_lat))
# Set the original CRS to WGS84 (lat/lon)
gdf.crs = 'EPSG:4326'  # WGS 84
# Reproject the GeoDataFrame to WGS 84 / UTM zone 32N
gdf = gdf.to_crs(epsg=32632)
# Select the fields you want to include in the shapefile
fields = ['stop_id', 'stop_name', 'stop_desc', 'pums', 's_lat', 's_lon', 'int_scr', 'scr_abc','Cluster', 'catch_area']
gdf = gdf[fields]
gdf.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\02-Catchment_area\02-metro_catchment.csv")

### 2.3 Tram

gdf=tram_scr

# Define the assign_
def assign_cluster(score, low_threshold, high_threshold):
    if score <= low_threshold:
        return 'Low'
    elif score > low_threshold and score <= high_threshold:
        return 'Medium'
    else:
        return 'High'

# Assuming the 'catch_area' column exists and is already calculated as 'scr_abc' * 35
gdf['catch_area'] = gdf['scr_abc'] * 35

# Calculate the thresholds for clustering
low_threshold = gdf['scr_abc'].quantile(0.25)
high_threshold = gdf['scr_abc'].quantile(0.75)

# Apply the function to create a new column 'Cluster'
gdf['Cluster'] = gdf['scr_abc'].apply(lambda x: assign_cluster(x, low_threshold, high_threshold))

# Find the maximum 'catch_area' for each cluster
max_catch_area_per_cluster = gdf.groupby('Cluster')['catch_area'].transform('max')

# Assign the max 'catch_area' to all members of the same cluster
gdf['catch_area'] = max_catch_area_per_cluster
gdf = gpd.GeoDataFrame(gdf, geometry=gpd.points_from_xy(gdf.s_lon, gdf.s_lat))
# Set the original CRS to WGS84 (lat/lon)
gdf.crs = 'EPSG:4326'  # WGS 84
# Reproject the GeoDataFrame to WGS 84 / UTM zone 32N
gdf = gdf.to_crs(epsg=32632)
# Select the fields you want to include in the shapefile
fields = ['stop_id', 'stop_name', 'stop_desc', 'pums', 's_lat', 's_lon', 'int_scr', 'scr_abc','Cluster', 'catch_area']
gdf = gdf[fields]
gdf.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\02-Catchment_area\02-tram_catchment.csv")

### 2.4 Bus

gdf=bus_scr

# Define the assign_
def assign_cluster(score, low_threshold, high_threshold):
    if score <= low_threshold:
        return 'Low'
    elif score > low_threshold and score <= high_threshold:
        return 'Medium'
    else:
        return 'High'

# Assuming the 'catch_area' column exists and is already calculated as 'scr_abc' * 35
gdf['catch_area'] = gdf['scr_abc'] * 35

# Calculate the thresholds for clustering
low_threshold = gdf['scr_abc'].quantile(0.25)
high_threshold = gdf['scr_abc'].quantile(0.75)

# Apply the function to create a new column 'Cluster'
gdf['Cluster'] = gdf['scr_abc'].apply(lambda x: assign_cluster(x, low_threshold, high_threshold))

# Find the maximum 'catch_area' for each cluster
max_catch_area_per_cluster = gdf.groupby('Cluster')['catch_area'].transform('max')

# Assign the max 'catch_area' to all members of the same cluster
gdf['catch_area'] = max_catch_area_per_cluster
gdf = gpd.GeoDataFrame(gdf, geometry=gpd.points_from_xy(gdf.s_lon, gdf.s_lat))
# Set the original CRS to WGS84 (lat/lon)
gdf.crs = 'EPSG:4326'  # WGS 84
# Reproject the GeoDataFrame to WGS 84 / UTM zone 32N
gdf = gdf.to_crs(epsg=32632)
# Select the fields you want to include in the shapefile
fields = ['stop_id', 'stop_name', 'stop_desc', 's_lat', 's_lon', 'int_scr', 'scr_abc','Cluster', 'catch_area']
gdf = gdf[fields]
gdf.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\02-Catchment_area\02-bus_catchment.csv")
