==> 02-01 Scoring Cleaning Duplicates.ipynb <==
import pandas as pd

# Load the dataset
bus = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\02-Catchment_area\02-bus_catchment.csv")
tram = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\02-Catchment_area\02-tram_catchment.csv")

# BUS

data = bus

# Group data by 'stop_name' and 'scr_abc' to identify identical rows
grouped = data.groupby(['stop_name', 'scr_abc'])

# Initialize a list to hold the processed rows
processed_rows = []

for _, group in grouped:
    if len(group) > 1:  # Check if there are identical rows
        # Calculate the midpoint of s_lat and s_lon for the group
        mid_s_lat = group['s_lat'].mean()
        mid_s_lon = group['s_lon'].mean()

        # Keep the first record's parameters for other fields
        first_record = group.iloc[0].copy()
        first_record['s_lat'] = mid_s_lat
        first_record['s_lon'] = mid_s_lon

        processed_rows.append(first_record)
    else:
        # If there's only one row in the group, just append it
        processed_rows.append(group.iloc[0])

# Create a new DataFrame from the processed rows
processed_data = pd.DataFrame(processed_rows).reset_index(drop=True)
processed_data.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\02-Catchment_area\02-01-bus_catchment.csv")

# TRAM

data = tram

# Group data by 'stop_name' and 'scr_abc' to identify identical rows
grouped = data.groupby(['stop_name', 'scr_abc'])

# Initialize a list to hold the processed rows
processed_rows = []

processed_rows = []

for _, group in grouped:
    # Skip the group if 'stop_name' is NaN or if 'pums' is 'yes'
    if pd.isna(group['stop_name'].iloc[0]) or (group['pums'].iloc[0].lower() == 'yes'):
        processed_rows.append(group.iloc[0])
        continue
    
    if len(group) > 1:
        # Calculate the midpoint of Lat and Lon for the group
        mid_s_lat = group['s_lat'].mean()
        mid_s_lon = group['s_lon'].mean()

        # Keep the first record's parameters for other fields
        first_record = group.iloc[0].copy()
        first_record['s_lat'] = mid_s_lat
        first_record['s_lon'] = mid_s_lon

        processed_rows.append(first_record)
    else:
        # If there's only one row in the group, just append it
        processed_rows.append(group.iloc[0])

# Create a new DataFrame from the processed rows
processed_data = pd.DataFrame(processed_rows).reset_index(drop=True)


# Create a new DataFrame from the processed rows
processed_data = pd.DataFrame(processed_rows).reset_index(drop=True)
processed_data.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\02-Catchment_area\02-01-tram_catchment1.csv")

processed_data
