==> 03-Merge data .ipynb <==
import pandas as pd

# BUS

n01_bus = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_01\outputs\n01-bus.csv")
n02_bus = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_02\outputs\n02-bus.csv")
n03_bus = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_03\outputs\n03-bus.csv")
n04_bus = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_04\output\n04-bus.csv")
n05_bus = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_05\outputs\n05-bus.csv")

# Start with the first DataFrame and merge the others one by one on 'stop_id'
merged_df = pd.merge(n01_bus, n02_bus[['stop_id', 'N2_1', 'N2_2']], on='stop_id', how='outer')
merged_df = pd.merge(merged_df, n03_bus[['stop_id', 'N3_1', 'N3_2', 'N3_3']], on='stop_id', how='outer')
merged_df = pd.merge(merged_df, n04_bus[['stop_id', 'N4_1', 'N4_2']], on='stop_id', how='outer')
merged_df = pd.merge(merged_df, n05_bus[['stop_id', 'n5_1', 'n5_2']], on='stop_id', how='outer')


merged_df.drop(['Unnamed: 0', 'field_1'], axis=1, inplace=True, errors='ignore')

merged_df.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\03-Merged\bus-pums.csv", index=False)

#  Train

n01_train = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_01\outputs\n01-train.csv")
n02_train = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_02\outputs\n02-train.csv")
n03_train = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_03\outputs\n03-train.csv")
n04_train = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_04\output\n04-train.csv")
n05_train = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_05\outputs\n05-train.csv")

# Start with the first DataFrame and merge the others one by one on 'stop_id'
merged_df = pd.merge(n01_train, n02_train[['field_1', 'N2_1', 'N2_2']], on='field_1', how='outer')
merged_df = pd.merge(merged_df, n03_train[['field_1', 'N3_1', 'N3_2', 'N3_3']], on='field_1', how='outer')
merged_df = pd.merge(merged_df, n04_train[['field_1', 'N4_1', 'N4_2']], on='field_1', how='outer')
merged_df = pd.merge(merged_df, n05_train[['field_1', 'N5_1', 'N5_2']], on='field_1', how='outer')

merged_df.drop(['Unnamed: 0', 'field_1'], axis=1, inplace=True, errors='ignore')

merged_df.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\03-Merged\train-pums.csv", index=False)

# METRO

n01_metro = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_01\outputs\n01-metro.csv")
n02_metro = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_02\outputs\n02-metro.csv")
n03_metro = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_03\outputs\n03-metro.csv")
n04_metro = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_04\output\n04-metro.csv")
n05_metro = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_05\outputs\n05-metro.csv")

# Start with the first DataFrame and merge the others one by one on 'stop_id'
merged_df = pd.merge(n01_metro, n02_metro[['field_1', 'N2_1', 'N2_2']], on='field_1', how='outer')
merged_df = pd.merge(merged_df, n03_metro[['field_1', 'N3_1', 'N3_2', 'N3_3']], on='field_1', how='outer')
merged_df = pd.merge(merged_df, n04_metro[['field_1', 'N4_1', 'N4_2']], on='field_1', how='outer')
merged_df = pd.merge(merged_df, n05_metro[['field_1', 'N5_1', 'N5_2']], on='field_1', how='outer')

merged_df.drop(['Unnamed: 0', 'field_1'], axis=1, inplace=True, errors='ignore')

merged_df.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\03-Merged\metro-pums.csv", index=False)

# TRAM

n01_tram = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_01\outputs\n01-tram.csv")
n02_tram = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_02\outputs\n02-tram.csv")
n03_tram = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_03\outputs\n03-tram.csv")
n04_tram = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_04\output\n04-tram.csv")
n05_tram = pd.read_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\n_05\outputs\n05-tram.csv")

# Start with the first DataFrame and merge the others one by one on 'stop_id'
merged_df = pd.merge(n01_tram, n02_tram[['s_lat', 'N2_1', 'N2_2']], on='s_lat', how='outer')
merged_df = pd.merge(merged_df, n03_tram[['s_lat', 'N3_1', 'N3_2', 'N3_3']], on='s_lat', how='outer')
merged_df = pd.merge(merged_df, n04_tram[['s_lat', 'N4_1', 'N4_2']], on='s_lat', how='outer')
merged_df = pd.merge(merged_df, n05_tram[['s_lat', 'N5_1', 'N5_2']], on='s_lat', how='outer')

merged_df.drop(['Unnamed: 0', 'field_1'], axis=1, inplace=True, errors='ignore')

merged_df.to_csv(r"D:\15_Polimi\Polimi_onedrive\OneDrive - Politecnico di Milano\PhD\Mapping\ABC\ABC-model-18-02-2024\results\03-Merged\tram-pums.csv", index=False)

merged_df
